+++ src/SBBox.cpp
@@ -22,9 +22,16 @@
 #include "SBBox.h"
 #include "SBBoxImpl.h"
 #include "FFT.h"
+#include "Interpolant.h"  // For sinc(x)
+
+// cf. comments about USE_COS_SIN in SBGaussian.cpp
+#ifdef _INTEL_COMPILER
+#define USE_COS_SIN
+#endif
 
 #ifdef DEBUGLOGGING
 #include <fstream>
+//std::ostream* dbgout = &std::cerr;
 //std::ostream* dbgout = new std::ofstream("debug.out");
 //int verbose_level = 2;
 #endif
@@ -57,39 +64,73 @@ namespace galsim {
     {
         if (_height==0.) _height=_width;
         _norm = _flux / (_width * _height);
+        _wo2 = 0.5*_width;
+        _ho2 = 0.5*_height;
+        _wo2pi = _width/(2.*M_PI);
+        _ho2pi = _height/(2.*M_PI);
     }
 
 
     double SBBox::SBBoxImpl::xValue(const Position<double>& p) const 
     {
-        if (fabs(p.x) < 0.5*_width && fabs(p.y) < 0.5*_height) return _norm;
+        if (fabs(p.x) < _wo2 && fabs(p.y) < _ho2) return _norm;
         else return 0.;  // do not use this function for filling image!
-    }
-
-    double SBBox::SBBoxImpl::sinc(double u) const 
-    {
-        if (std::abs(u) < 1.e-3)
-            return 1.-u*u/6.;
-        else
-            return std::sin(u)/u;
     }
 
     std::complex<double> SBBox::SBBoxImpl::kValue(const Position<double>& k) const
     {
-        return _flux * sinc(0.5*k.x*_width)*sinc(0.5*k.y*_height);
+        return _flux * sinc(k.x*_wo2pi)*sinc(k.y*_ho2pi);
     }
 
     void SBBox::SBBoxImpl::fillXValue(tmv::MatrixView<double> val,
-                                      double x0, double dx, int ix_zero,
-                                      double y0, double dy, int iy_zero) const
+                                      double x0, double dx, int izero,
+                                      double y0, double dy, int jzero) const
     {
         dbg<<"SBBox fillXValue\n";
-        dbg<<"x = "<<x0<<" + ix * "<<dx<<", ix_zero = "<<ix_zero<<std::endl;
-        dbg<<"y = "<<y0<<" + iy * "<<dy<<", iy_zero = "<<iy_zero<<std::endl;
+        dbg<<"x = "<<x0<<" + i * "<<dx<<", izero = "<<izero<<std::endl;
+        dbg<<"y = "<<y0<<" + j * "<<dy<<", jzero = "<<jzero<<std::endl;
+
         assert(val.stepi() == 1);
         const int m = val.colsize();
         const int n = val.rowsize();
         typedef tmv::VIt<double,1,tmv::NonConj> It;
+
+        // It will be useful to do everything in units of dx,dy
+        x0 /= dx;
+        double wo2 = _wo2 / dx;
+        y0 /= dy;
+        double ho2 = _ho2 / dy;
+        xdbg<<"x0,y0 -> "<<x0<<','<<y0<<std::endl;
+        xdbg<<"width,height -> "<<wo2*2.<<','<<ho2*2.<<std::endl;
+
+        // Start by setting everything to zero
+        val.setZero();
+
+        // Then fill the interior with _norm:
+        // Fill pixels where:
+        //     x0 + ix >= -width/2
+        //     x0 + ix < width/2
+        //     y0 + iy >= -width/2
+        //     y0 + iy < width/2
+
+        int ix1 = std::max(0, int(std::ceil(-wo2 - x0)));
+        int ix2 = std::min(m, int(std::ceil(wo2 - x0)));
+        int iy1 = std::max(0, int(std::ceil(-ho2 - y0)));
+        int iy2 = std::min(n, int(std::ceil(ho2 - y0)));
+
+        if (ix1 < ix2 && iy1 < iy2)
+            val.subMatrix(ix1,ix2,iy1,iy2).setAllTo(_norm);
+
+#if 0
+        // We used to implement this by making the pixels that cross the edge have a 
+        // fractional flux value appropriate for the fraction of the box that goes through
+        // each pixel.  However, this isn't actually correct.  SBProfile objects are always
+        // rendered as the local surface brightness at the center of the pixel.  To get
+        // the right flux, you need to convolve by a Pixel.  So if someone renders a Box
+        // without convolving by a pixel, it is inconsistent to do this differently than we
+        // do all the other SBProfile types.  However, since it was an involved calculation
+        // and someone might actually care to resurrect it in a different guise at some point,
+        // I'm leaving it here, just commented out.
 
         // We need to make sure the pixels where the edges of the box fall only get
         // a fraction of the flux.
@@ -105,14 +146,6 @@ namespace galsim {
         //    above the box where val = 0 again
         //
         // Furthermore, we have to calculate the correct values for the pixels on the border.
-        
-        // It will be useful to do everything in units of dx,dy
-        x0 /= dx;
-        double width = _width / dx;
-        y0 /= dy;
-        double height = _height / dy;
-        xdbg<<"x0,y0 -> "<<x0<<','<<y0<<std::endl;
-        xdbg<<"width,height -> "<<width<<','<<height<<std::endl;
 
         int ix_left, ix_right, iy_bottom, iy_top;
         double x_left, x_right, y_bottom, y_top;
@@ -182,18 +215,19 @@ namespace galsim {
         val(ix_right,iy_bottom) = x_right * y_bottom * _norm;
         val(ix_left,iy_top) = x_left * y_top * _norm;
         val(ix_right,iy_top) = x_right * y_top * _norm;
+#endif
     }
 
     void SBBox::SBBoxImpl::fillKValue(tmv::MatrixView<std::complex<double> > val,
-                                      double x0, double dx, int ix_zero,
-                                      double y0, double dy, int iy_zero) const
+                                      double kx0, double dkx, int izero,
+                                      double ky0, double dky, int jzero) const
     {
         dbg<<"SBBox fillKValue\n";
-        dbg<<"x = "<<x0<<" + ix * "<<dx<<", ix_zero = "<<ix_zero<<std::endl;
-        dbg<<"y = "<<y0<<" + iy * "<<dy<<", iy_zero = "<<iy_zero<<std::endl;
-        if (ix_zero != 0 || iy_zero != 0) {
+        dbg<<"kx = "<<kx0<<" + i * "<<dkx<<", izero = "<<izero<<std::endl;
+        dbg<<"ky = "<<ky0<<" + j * "<<dky<<", jzero = "<<jzero<<std::endl;
+        if (izero != 0 || jzero != 0) {
             xdbg<<"Use Quadrant\n";
-            fillKValueQuadrant(val,x0,dx,ix_zero,y0,dy,iy_zero);
+            fillKValueQuadrant(val,kx0,dkx,izero,ky0,dky,jzero);
         } else {
             xdbg<<"Non-Quadrant\n";
             assert(val.stepi() == 1);
@@ -201,21 +235,21 @@ namespace galsim {
             const int n = val.rowsize();
             typedef tmv::VIt<double,1,tmv::NonConj> It;
 
-            x0 *= 0.5*_width;
-            dx *= 0.5*_width;
-            y0 *= 0.5*_height;
-            dy *= 0.5*_height;
+            kx0 *= _wo2pi;
+            dkx *= _wo2pi;
+            ky0 *= _ho2pi;
+            dky *= _ho2pi;
 
             // The Box profile in Fourier space is separable:
-            //    val = _flux * sinc(0.5 * x * _height) * sinc(0.5 * y * _height) 
-            tmv::Vector<double> sinc_x(m);
-            It xit = sinc_x.begin();
-            for (int i=0;i<m;++i,x0+=dx) *xit++ = sinc(x0);
-            tmv::Vector<double> sinc_y(n);
-            It yit = sinc_y.begin();
-            for (int j=0;j<n;++j,y0+=dy) *yit++ = sinc(y0);
+            //    val(x,y) = _flux * sinc(x * _width/2pi) * sinc(y * _height/2pi) 
+            tmv::Vector<double> sinc_kx(m);
+            It kxit = sinc_kx.begin();
+            for (int i=0;i<m;++i,kx0+=dkx) *kxit++ = sinc(kx0);
+            tmv::Vector<double> sinc_ky(n);
+            It kyit = sinc_ky.begin();
+            for (int j=0;j<n;++j,ky0+=dky) *kyit++ = sinc(ky0);
 
-            val = _flux * sinc_x ^ sinc_y;
+            val = _flux * sinc_kx ^ sinc_ky;
         }
     }
 
@@ -223,44 +257,54 @@ namespace galsim {
                                       double x0, double dx, double dxy,
                                       double y0, double dy, double dyx) const
     {
-        // This is complicated to get right, since the edges cut through the image grid
-        // at angles.  Fortunately, we also don't really have any need for it.
-        // It would only get called if you draw a sheared or rotated Pixel without convolving 
-        // it by anything else, which we don't really do.  If we ever decide we need this,
-        // someone can try to figure out all the math involved here.
-        if (dxy == 0. && dyx == 0.)
-            fillXValue(val,x0,dx,0,y0,dy,0);
-        else
-            throw std::runtime_error(
-                "fillXValue for a sheared or rotated SBBox is not implemented.");
+        dbg<<"SBBox fillXValue\n";
+        dbg<<"x = "<<x0<<" + i * "<<dx<<" + j * "<<dxy<<std::endl;
+        dbg<<"y = "<<y0<<" + i * "<<dyx<<" + j * "<<dy<<std::endl;
+        assert(val.stepi() == 1);
+        assert(val.canLinearize());
+        const int m = val.colsize();
+        const int n = val.rowsize();
+        typedef tmv::VIt<double,1,tmv::NonConj> It;
+
+        It valit = val.linearView().begin();
+        for (int j=0;j<n;++j,x0+=dxy,y0+=dy) {
+            double x = x0;
+            double y = y0;
+            int i=0;
+            // Use the fact that any slice through the box has only one segment that is non-zero.
+            // So start with zeroes until in the box, then _norm, then more zeroes.
+            for (;i<m && (std::abs(x)>_wo2 || std::abs(y)>_ho2); ++i,x+=dx,y+=dyx) *valit++ = 0.;
+            for (;i<m && std::abs(x)<_wo2 && std::abs(y)<_ho2; ++i,x+=dx,y+=dyx) *valit++ = _norm;
+            for (;i<m; ++i,x+=dx,y+=dyx) *valit++ = 0.;
+        }
     }
 
     void SBBox::SBBoxImpl::fillKValue(tmv::MatrixView<std::complex<double> > val,
-                                      double x0, double dx, double dxy,
-                                      double y0, double dy, double dyx) const
+                                      double kx0, double dkx, double dkxy,
+                                      double ky0, double dky, double dkyx) const
     {
         dbg<<"SBBox fillKValue\n";
-        dbg<<"x = "<<x0<<" + ix * "<<dx<<" + iy * "<<dxy<<std::endl;
-        dbg<<"y = "<<y0<<" + ix * "<<dyx<<" + iy * "<<dy<<std::endl;
+        dbg<<"kx = "<<kx0<<" + i * "<<dkx<<" + j * "<<dkxy<<std::endl;
+        dbg<<"ky = "<<ky0<<" + i * "<<dkyx<<" + j * "<<dky<<std::endl;
         assert(val.stepi() == 1);
         assert(val.canLinearize());
         const int m = val.colsize();
         const int n = val.rowsize();
         typedef tmv::VIt<std::complex<double>,1,tmv::NonConj> It;
 
-        x0 *= 0.5*_width;
-        dx *= 0.5*_width;
-        dxy *= 0.5*_width;
-        y0 *= 0.5*_height;
-        dy *= 0.5*_height;
-        dyx *= 0.5*_height;
+        kx0 *= _wo2pi;
+        dkx *= _wo2pi;
+        dkxy *= _wo2pi;
+        ky0 *= _ho2pi;
+        dky *= _ho2pi;
+        dkyx *= _ho2pi;
 
-        It valit(val.linearView().begin().getP(),1);
-        for (int j=0;j<n;++j,x0+=dxy,y0+=dy) {
-            double x = x0;
-            double y = y0;
-            for (int i=0;i<m;++i,x+=dx,y+=dyx) 
-                *valit++ = _flux * sinc(x) * sinc(y);
+        It valit = val.linearView().begin();
+        for (int j=0;j<n;++j,kx0+=dkxy,ky0+=dky) {
+            double kx = kx0;
+            double ky = ky0;
+            for (int i=0;i<m;++i,kx+=dkx,ky+=dkyx) 
+                *valit++ = _flux * sinc(kx) * sinc(ky);
         }
     }
 
@@ -283,9 +327,239 @@ namespace galsim {
         dbg<<"Box shoot: N = "<<N<<std::endl;
         dbg<<"Target flux = "<<getFlux()<<std::endl;
         boost::shared_ptr<PhotonArray> result(new PhotonArray(N));
-        for (int i=0; i<result->size(); i++)
-            result->setPhoton(i, _width*(u()-0.5), _height*(u()-0.5), _flux/N);
+        double fluxPerPhoton = _flux/N;
+        for (int i=0; i<N; i++)
+            result->setPhoton(i, _width*(u()-0.5), _height*(u()-0.5), fluxPerPhoton);
         dbg<<"Box Realized flux = "<<result->getTotalFlux()<<std::endl;
+        return result;
+    }
+
+
+
+    SBTopHat::SBTopHat(double radius, double flux, const GSParamsPtr& gsparams) :
+        SBProfile(new SBTopHatImpl(radius,flux,gsparams)) {}
+
+    SBTopHat::SBTopHat(const SBTopHat& rhs) : SBProfile(rhs) {}
+
+    SBTopHat::~SBTopHat() {}
+
+    double SBTopHat::getRadius() const 
+    {
+        assert(dynamic_cast<const SBTopHatImpl*>(_pimpl.get()));
+        return static_cast<const SBTopHatImpl&>(*_pimpl).getRadius(); 
+    }
+
+    SBTopHat::SBTopHatImpl::SBTopHatImpl(double radius, double flux,
+                                         const GSParamsPtr& gsparams) :
+        SBProfileImpl(gsparams),
+        _r0(radius), _r0sq(_r0*_r0), _flux(flux),
+        _norm(_flux / (M_PI * _r0sq))
+    {
+    }
+
+
+    double SBTopHat::SBTopHatImpl::xValue(const Position<double>& p) const 
+    {
+        double rsq = p.x*p.x + p.y*p.y;
+        if (rsq < _r0sq) return _norm;
+        else return 0.;
+    }
+
+    std::complex<double> SBTopHat::SBTopHatImpl::kValue(const Position<double>& k) const
+    {
+        double kr0sq = (k.x*k.x + k.y*k.y) * _r0sq;
+        return kValue2(kr0sq);
+    }
+
+    std::complex<double> SBTopHat::SBTopHatImpl::kValue2(double kr0sq) const
+    {
+        if (kr0sq < 1.e-4) {
+            // Use the Taylor expansion for small arguments.
+            // Error from omitting next term is about 1.e-16 for kr0sq = 1.e-4
+            return _flux * (1. - kr0sq * ( (1./8.) + (1./192.) * kr0sq ));
+        } else {
+            double kr0 = sqrt(kr0sq);
+            return 2.*_flux * j1(kr0)/kr0;
+        }
+    }
+
+    void SBTopHat::SBTopHatImpl::fillXValue(tmv::MatrixView<double> val,
+                                            double x0, double dx, int izero,
+                                            double y0, double dy, int jzero) const
+    {
+        dbg<<"SBTopHat fillXValue\n";
+        dbg<<"x = "<<x0<<" + i * "<<dx<<", izero = "<<izero<<std::endl;
+        dbg<<"y = "<<y0<<" + j * "<<dy<<", jzero = "<<jzero<<std::endl;
+
+        assert(val.stepi() == 1);
+        const int m = val.colsize();
+        const int n = val.rowsize();
+        typedef tmv::VIt<double,1,tmv::NonConj> It;
+
+        val.setZero();
+        // The columns to consider have -r0 <= y < r0
+        // given that y = y0 + j dy
+        int j1 = std::max(0, int(std::ceil((-_r0 - y0)/dy)));
+        int j2 = std::min(n, int(std::ceil((_r0 - y0)/dy)));
+        y0 += j1 * dy;
+        for (int j=j1;j<j2;++j,y0+=dy) {
+            double ysq = y0*y0;
+            double xmax = std::sqrt(_r0sq - ysq);
+            // Set to _norm all pixels with -xmax <= x < xmax
+            // given that x = x0 + i dx.
+            int i1 = std::max(0, int(std::ceil((-xmax - x0)/dx)));
+            int i2 = std::min(m, int(std::ceil((xmax - x0)/dx)));
+            if (i1 < i2)
+                val.col(j,i1,i2).setAllTo(_norm);
+        }
+    }
+
+    void SBTopHat::SBTopHatImpl::fillKValue(tmv::MatrixView<std::complex<double> > val,
+                                            double kx0, double dkx, int izero,
+                                            double ky0, double dky, int jzero) const
+    {
+        dbg<<"SBTopHat fillKValue\n";
+        dbg<<"kx = "<<kx0<<" + i * "<<dkx<<", izero = "<<izero<<std::endl;
+        dbg<<"ky = "<<ky0<<" + j * "<<dky<<", jzero = "<<jzero<<std::endl;
+        if (izero != 0 || jzero != 0) {
+            xdbg<<"Use Quadrant\n";
+            fillKValueQuadrant(val,kx0,dkx,izero,ky0,dky,jzero);
+        } else {
+            xdbg<<"Non-Quadrant\n";
+            assert(val.stepi() == 1);
+            const int m = val.colsize();
+            const int n = val.rowsize();
+            typedef tmv::VIt<std::complex<double>,1,tmv::NonConj> It;
+
+            kx0 *= _r0;
+            dkx *= _r0;
+            ky0 *= _r0;
+            dky *= _r0;
+
+            for (int j=0;j<n;++j,ky0+=dky) {
+                double kx = kx0;
+                double kysq = ky0*ky0;
+                It valit = val.col(j).begin();
+                for (int i=0;i<m;++i,kx+=dkx) {
+                    double ksq = kx*kx + kysq;
+                    *valit++ = kValue2(ksq);
+                }
+            }
+        }
+    }
+
+    void SBTopHat::SBTopHatImpl::fillXValue(tmv::MatrixView<double> val,
+                                            double x0, double dx, double dxy,
+                                            double y0, double dy, double dyx) const
+    {
+        dbg<<"SBTopHat fillXValue\n";
+        dbg<<"x = "<<x0<<" + i * "<<dx<<" + j * "<<dxy<<std::endl;
+        dbg<<"y = "<<y0<<" + i * "<<dyx<<" + j * "<<dy<<std::endl;
+        assert(val.stepi() == 1);
+        assert(val.canLinearize());
+        const int m = val.colsize();
+        const int n = val.rowsize();
+        typedef tmv::VIt<double,1,tmv::NonConj> It;
+
+        val.setZero();
+        It valit = val.linearView().begin();
+        for (int j=0;j<n;++j,x0+=dxy,y0+=dy) {
+            double x = x0;
+            double y = y0;
+            int i=0;
+            // Use the fact that any slice through the circle has only one segment that is non-zero.
+            // So start with zeroes until in the circle, then _norm, then more zeroes.
+            // Note: this could be sped up somewhat using the same kind of calculation we did
+            // for the non-sheared fillXValue (the one with izero, jzero), but I didn't
+            // bother.  This is probably plenty fast enough for as often as the function is 
+            // called (i.e. almost never!)
+            for (;i<m && (x*x+y*y > _r0sq); ++i,x+=dx,y+=dyx) ++valit;
+            for (;i<m && (x*x+y*y < _r0sq); ++i,x+=dx,y+=dyx) *valit++ = _norm;
+            for (;i<m; ++i,x+=dx,y+=dyx) ++valit;
+        }
+    }
+
+    void SBTopHat::SBTopHatImpl::fillKValue(tmv::MatrixView<std::complex<double> > val,
+                                            double kx0, double dkx, double dkxy,
+                                            double ky0, double dky, double dkyx) const
+    {
+        dbg<<"SBTopHat fillKValue\n";
+        dbg<<"kx = "<<kx0<<" + i * "<<dkx<<" + j * "<<dkxy<<std::endl;
+        dbg<<"ky = "<<ky0<<" + i * "<<dkyx<<" + j * "<<dky<<std::endl;
+        assert(val.stepi() == 1);
+        assert(val.canLinearize());
+        const int m = val.colsize();
+        const int n = val.rowsize();
+        typedef tmv::VIt<std::complex<double>,1,tmv::NonConj> It;
+
+        kx0 *= _r0;
+        dkx *= _r0;
+        dkxy *= _r0;
+        ky0 *= _r0;
+        dky *= _r0;
+        dkyx *= _r0;
+
+        It valit = val.linearView().begin();
+        for (int j=0;j<n;++j,kx0+=dkxy,ky0+=dky) {
+            double kx = kx0;
+            double ky = ky0;
+            for (int i=0;i<m;++i,kx+=dkx,ky+=dkyx) {
+                double ksq = kx*kx + ky*ky;
+                *valit++ = kValue2(ksq);
+            }
+        }
+    }
+
+    // Set maxK to the value where the FT is down to maxk_threshold
+    double SBTopHat::SBTopHatImpl::maxK() const 
+    { 
+        // |j1(x)| ~ sqrt(2/(Pi x)) for large x, so using this, we get
+        // maxk_thresh = 2 * sqrt(2/(Pi k r0)) / (k r0) = 2 sqrt(2/Pi) (k r0)^-3/2
+        return std::pow(2. * sqrt(2./M_PI) / this->gsparams->maxk_threshold, 2./3.) / _r0;
+    }
+
+    // The amount of flux missed in a circle of radius pi/stepk should be at 
+    // most folding_threshold of the flux.
+    double SBTopHat::SBTopHatImpl::stepK() const
+    {
+        // _r0 encloses all the flux, so use that.
+        return M_PI / _r0;
+    }
+
+    boost::shared_ptr<PhotonArray> SBTopHat::SBTopHatImpl::shoot(int N, UniformDeviate u) const
+    {
+        dbg<<"TopHat shoot: N = "<<N<<std::endl;
+        dbg<<"Target flux = "<<getFlux()<<std::endl;
+        boost::shared_ptr<PhotonArray> result(new PhotonArray(N));
+        double fluxPerPhoton = _flux/N;
+        // cf. SBGaussian's shoot function
+        for (int i=0; i<N; i++) {
+            // First get a point uniformly distributed on unit circle
+#ifdef USE_COS_SIN
+            double theta = 2.*M_PI*u();
+            double rsq = u(); // cumulative dist function P(<r) = r^2 for unit circle
+#ifdef _GLIBCXX_HAVE_SINCOS
+            // Most optimizing compilers will do this automatically, but just in case...
+            double sint,cost;
+            sincos(theta,&sint,&cost);
+#else
+            double cost = std::cos(theta);
+            double sint = std::sin(theta);
+#endif
+            // Then map radius to the desired Gaussian with analytic transformation
+            double r = sqrt(rsq) * _r0;;
+            result->setPhoton(i, r*cost, r*sint, fluxPerPhoton);
+#else
+            double xu, yu, rsq;
+            do {
+                xu = 2.*u()-1.;
+                yu = 2.*u()-1.;
+                rsq = xu*xu+yu*yu;
+            } while (rsq>=1.);
+            result->setPhoton(i, xu * _r0, yu * _r0, fluxPerPhoton);
+#endif
+        }
+        dbg<<"TopHat Realized flux = "<<result->getTotalFlux()<<std::endl;
         return result;
     }
 }